Fix GCC CI: LuaVal init block and sign-compare warnings.
PrepareClientAddons returns size_t; cast addon count for LuaVal. Add shared AIO.h and silence -Wsign-compare in AIO paths.

+++ src/server/game/Chat/AIOMsg.cpp
@@ -61,7 +61,7 @@ AIOMsg& AIOMsg::AppendLast(LuaVal const& a1, LuaVal const& a2, LuaVal const& a3,
     if (!nArgsVal.isnumber())
         return *this;
 
-    unsigned int nArgs = nArgsVal.num();
+    unsigned int nArgs = static_cast<unsigned int>(nArgsVal.num());
     if (!a1.isnil())
     {
         block.insert(a1);
